test(compiler-plugin/previewCollection/ir): split IR tests by logic (collectPreviewsReplacement / buildPreviewSequence)
Ticket 1 で確定した production の IR logic 構造に sync する。本 commit では IR 側の
plugin 挙動を検証するテスト群を logic 別 (collectPreviewsReplacement /
buildPreviewSequence sub-logic) に再配置する。

`ir/collectPreviewsReplacement/` (`collect[All]ModulePreviews()` の置換 +
依存モジュールの hint discovery + hint body 埋め込み):

- `PreviewHintDiscoveryTest.kt` (consumer-side hint discovery)
- `PreviewHintIgnoreCrossModuleTest.kt` (`ignored` の cross-module behavior)
- `HintHashCollisionTest.kt` (`HashMapWithCollisionDetection.kt` の挙動)
- `HintCrossArtifactAndSquattingTest.kt` (`DiscoverHints.kt` の squatting /
  cross-artifact 検出 warning。Ticket 4 で warning 構造化後に literal
  assertion の追従更新が必要)

`ir/collectPreviewsReplacement/buildPreviewSequence/` (Sequence 構築 sub-logic =
`lazyPreviewSequence` + factory lambdas + lazy + PreviewExport + cross-module 連結 +
source/KDoc 抽出):

- `SourceTextExtractionTest.kt` (`ExtractedSourceText.kt` の挙動)

package: `feature` → `feature.previewCollection.ir.collectPreviewsReplacement[.buildPreviewSequence]`

`PreviewHintIgnoreCrossModuleTest` は `PreviewHintTestSupport` の internal
extension (`previewHintFacadeFiles` / `previewHintMarkerFiles`) に依存するため
対応する import を追加。

検証:
- `:compiler-plugin:ktlintFormat` + `:compiler-plugin:ktlintCheck` pass
- `:compiler-plugin:test --rerun-tasks` pass (131 tests total)
- `apiCheck` pass
- `find compiler-plugin -name build/api -type d` で `compat-k*/build/api/` 以外ゼロ
- wildcard import 増加なし
